Move getting climate history to store
This includes a bit of refactoring:
- Introduce `view::data` crate for data types that are specifically for
  sending to the view.
- Introduce two ClimateObservation and ClimateMeasurement data types.
  One for use within the app and one for sending to the view.
